+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to
 - 🎨(frontend) Remove var usage instead of let/const
 - 📝(frontend) Update documentation
 
+## Fixed
+
+- 🐛(frontend) Add explicit React import to ensure JSX compatibility
+
 ## [3.2.1] - 2025-05-06
 
 ## Fixed